Bastion / App-Server configured for BLESS
Configuration for the bastion and app-server uses cloud-init.

Bastion is setup to call the BLESS lambda function on a successful login
via PAM. As PAM modules are called as root it should be possible to add
further configuration to lock down access to EC2 metadata and the lambda
function to only the root user.

By default users of BLESS have access to the app-server with the user
'engineer', this can be configured from var.bless_user. This is another
point where restrictions could be made. Currently the `bless_user` is
allowed passwordless sudo access on the app-server.

The actual BLESS lambda function isn't deployed by terraform so the arn
needs to be configured (var.bless_function). If possible it would be
good to deploy the lambda function using terraform but since it has
compiled dependencies it would need a build-phase which I don't think
should be implemented in terraform.